Moved ucdl-Tokens and renamed dto-package from parser to ucdl.
