+++ server/src/main/java/de/uftos/repositories/ucdl/parser/DefinitionParser.java
@@ -1,7 +1,7 @@
 package de.uftos.repositories.ucdl.parser;
 
-import de.uftos.dto.parser.AbstractSyntaxTreeDto;
-import de.uftos.dto.parser.ValueDto;
+import de.uftos.dto.ucdl.AbstractSyntaxTreeDto;
+import de.uftos.dto.ucdl.ValueDto;
 
 public class DefinitionParser {
   public static AbstractSyntaxTreeDto parseDefinition(String definition) {
